added markdown and json format output download
