+++ package.json
@@ -17,6 +17,7 @@
     "@vuelidate/validators": "^2.0.0-alpha.30",
     "axios": "^0.24.0",
     "core-js": "^3.6.5",
+    "lodash": "^4.17.21",
     "vue": "^3.0.0",
     "vue-router": "^4.0.0-0",
     "vuedraggable": "^4.1.0",
